refactor: Make types.py strictly typechecked.
We are now making the types module typecheck in strict mode.
This is mostly achieved by passing the correct types to the
generic type variables.

We ran into one specific issue with `JSONRPCRequest` and
`JSONRPCNotification`. Both are generic classes that take a
`dict[str, Any]` as params and just a plain string as method.

However, since the TypeVar for `RequestT` and `NotificationT` are
bound to `RequestParams` and `NotificationParams` respectively
we get into a type issue.

There are two ways of solving this:
1. Widen the bound by allowing explicitly for `dict[str, Any]`
2. Make JSONRPCRequest and JSONRPCNotificaiton not part of the
   type hierarchy with Request and Notification roots.

It felt most naturally to keep JSONRPCRequest/JSONRPCNotification
part of the type hierarchy and allow for general passing of
dict[str, Any].

This now typechecks.