Move uvicorn import to usage
uvicorn is an optional dependency only used in one function but is imported at the global scope meaning mcp imports will fail if uvicorn is not available.
This commit moves uvicorn import to usage in run_sse_async.
It would be ideal to mark uvicorn as a default-on optional dependency but that isn't supported yet by pyproject.toml.