Lazy import `jsonschema`
The import of `jsonschema` takes up almost 2s at cold start, which dramatically slow down the cold start in a containerized environment. Changing it to lazy import within the function that uses it will save those 2s at startup. As MCP is getting used more in production, this time-saving is critical.